# docs/api/openapi_v1.yaml
openapi: 3.0.3
info:
  title: Monitor Legislativo v4 API
  description: |
    Comprehensive API for monitoring Brazilian legislative documents and regulatory changes.
    
    ## Features
    - **Document Search**: Advanced search with filters and facets
    - **Real-time Monitoring**: Set up alerts for legislative changes
    - **Smart Alerts**: ML-powered relevance scoring
    - **Export Capabilities**: Multiple format support
    - **User Management**: Authentication and authorization
    
    ## Authentication
    This API uses JWT (JSON Web Token) authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to ensure fair usage:
    - **Authenticated users**: 1000 requests/hour
    - **Search endpoints**: 100 requests/minute
    - **Export endpoints**: 10 requests/minute
    
    ## Error Handling
    All endpoints return standardized error responses with structured error information.
    
  version: 1.0.0
  contact:
    name: Monitor Legislativo Support
    email: support@monitorlegislativo.com
    url: https://monitorlegislativo.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.monitorlegislativo.com/v1
    description: Production server
  - url: https://staging-api.monitorlegislativo.com/v1
    description: Staging server
  - url: http://localhost:5000/api/v1
...
security:
  - BearerAuth: []
paths:
...
  /alerts:
    get:
      tags:
        - Alerts
      summary: Get user alerts
      description: Retrieve monitoring alerts for the authenticated user
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, ACTIVE, RESOLVED, DISMISSED]
          description: Filter by alert status
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Filter by alert priority
      responses:
        '200':
          description: User alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Alerts
      summary: Create monitoring alert
      description: Create a new monitoring rule for legislative changes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertRequest'
      responses:
        '201':
          description: Alert created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /alerts/{alert_id}:
    get:
      tags:
        - Alerts
      summary: Get alert details
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: integer
          description: Alert ID
      responses:
        '200':
          description: Alert details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Alerts
      summary: Update alert
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: integer
          description: Alert ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlertRequest'
      responses:
        '200':
          description: Alert updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Alerts
      summary: Delete alert
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: integer
          description: Alert ID
      responses:
        '204':
          description: Alert deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'

  # Export Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
  schemas:
...
    StandardResponse:
      type: object
      required:
        - status
        - data
        - metadata
      properties:
        status:
          type: string
          enum: [success, error, warning, partial]
          description: Response status
        data:
          description: Response data (type varies by endpoint)
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: Error details (present when status is error)
        warnings:
          type: array
          items:
            type: string
          description: Warning messages
    ResponseMetadata:
      type: object
      required:
        - timestamp
        - version
        - request_id
        - execution_time_ms
      properties:
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        version:
          type: string
          description: API version
          example: "v1"
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
        execution_time_ms:
          type: number
          format: float
          description: Request execution time in milliseconds
        total_count:
          type: integer
          description: Total number of items (for paginated responses)
        page:
          type: integer
          description: Current page number
        per_page:
          type: integer
          description: Items per page
        has_more:
          type: boolean
          description: Whether more pages are available
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Query parameter is required"
        details:
          type: object
          description: Additional error context

    # Authentication Schemas
...
    Alert:
      type: object
      required:
        - id
        - title
        - status
        - created_at
      properties:
        id:
          type: integer
          description: Alert ID
        title:
          type: string
          description: Alert title
        message:
          type: string
          description: Alert message
        alert_type:
          type: string
          enum: [DOCUMENT_MATCH, KEYWORD_ALERT, STATUS_CHANGE, DEADLINE_REMINDER]
          description: Alert type
        status:
          type: string
          enum: [PENDING, ACTIVE, RESOLVED, DISMISSED]
          description: Alert status
        priority:
          type: string
          enum: [low, medium, high, critical]
          description: Alert priority
        relevance_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: ML-based relevance score
        is_read:
          type: boolean
          description: Whether alert has been read
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        read_at:
          type: string
          format: date-time
          description: Read timestamp
        expires_at:
          type: string
          format: date-time
          description: Expiration timestamp
    CreateAlertRequest:
      type: object
      required:
        - name
        - keywords
      properties:
        name:
          type: string
          maxLength: 200
          description: Alert name
          example: "LGPD Updates"
        description:
          type: string
          description: Alert description
        keywords:
          type: array
          items:
            type: string
          minItems: 1
          description: Keywords to monitor
          example: ["proteção de dados", "LGPD", "privacidade"]
        sources:
          type: array
          items:
            type: string
            enum: [Camara, Senado, Planalto, TCU, STF]
          description: Sources to monitor
        document_types:
          type: array
          items:
            type: string
            enum: [LEI, DECRETO, PORTARIA, RESOLUCAO, INSTRUCAO_NORMATIVA]
          description: Document types to monitor
        date_range_start:
          type: string
          format: date
          description: Start date for monitoring
        date_range_end:
          type: string
          format: date
          description: End date for monitoring
    UpdateAlertRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
        keywords:
          type: array
          items:
            type: string
        sources:
          type: array
          items:
            type: string
        document_types:
          type: array
          items:
            type: string
        is_enabled:
          type: boolean
    AlertResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Alert'
    AlertListResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                alerts:
                  type: array
                  items:
                    $ref: '#/components/schemas/Alert'

    # Export Schemas
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/StandardResponse'
              - type: object
                properties:
                  status:
                    type: string
                    enum: [error]
          example:
            status: error
            data: null
            metadata:
              timestamp: "2025-01-06T12:00:00Z"
              version: "v1"
              request_id: "123e4567-e89b-12d3-a456-426614174000"
              execution_time_ms: 45.2
            errors:
              - code: "VALIDATION_ERROR"
                message: "Query parameter is required"
                details:
                  field: "query"
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/StandardResponse'
              - type: object
                properties:
                  status:
                    type: string
                    enum: [error]
          example:
            status: error
            data: null
            metadata:
              timestamp: "2025-01-06T12:00:00Z"
              version: "v1"
              request_id: "123e4567-e89b-12d3-a456-426614174000"
              execution_time_ms: 12.1
            errors:
              - code: "AUTHENTICATION_ERROR"
                message: "Authentication required"
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/StandardResponse'
              - type: object
                properties:
                  status:
                    type: string
                    enum: [error]
...